Move all extensions into a concern and include that.
This does not change any behaviour, it just cleans out the class_eval
part.

Benefits:
 * ProductOnSalable is a cleaner place to store all the class-addons.
 * It becomes easier to test.
 * Users can easier override behaviour by re-opening our concern.
   Instead of re-opening a re-opend class.

E.g. one can now override the `put_on_sale` to add, say, a notifier
mailing out sale-newsletters like so:

Spree::ProductOnSalable.module_eval
  alias_method_chain :put_on_sale, :notifier
  def put_on_sale_with_notifier
    put_on_sale_without_notifier
    NewsletterMailer.new(:on_sale, self).deliver!
  end
end
